Add files via upload
Added config file support for changes and moved each option to its own class.
